+++ plugins/content/attachment/lib/api/filters/base/KalturaTextualAttachmentAssetBaseFilter.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @package plugins.attachment
+ * @relatedService AttachmentAssetService
+ * @subpackage api.filters.base
+ * @abstract
+ */
+abstract class KalturaTextualAttachmentAssetBaseFilter extends KalturaAttachmentAssetFilter
+{
+}