+++ plugins/content/markdown/lib/api/KalturaMarkdownAssetListResponse.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * @package plugins.markdown
+ * @subpackage api.objects
+ */
+class KalturaMarkdownAssetListResponse extends KalturaListResponse
+{
+	/**
+	 * @var KalturaMarkdownAssetArray
+	 * @readonly
+	 */
+	public $objects;
+}