+++ plugins/content/markdown/lib/api/enums/KalturaMarkdownProviderType.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * @package plugins.markdown
+ * @subpackage api.enum
+ */
+class KalturaMarkdownProviderType extends KalturaDynamicEnum implements MarkdownProviderType
+{
+	public static function getEnumClass()
+	{
+		return 'MarkdownProviderType';
+	}
+}